toolkit/components/passwordmgr/test/mochitest/test_usage_prefs.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test usage prefs</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: usage prefs collection

<script>
let DEFAULT_ORIGIN = window.location.origin;

function addLogin() {
  addLoginsInParent([DEFAULT_ORIGIN, DEFAULT_ORIGIN, null, "testuser", "testpass", "uname", "pword"]);
}

async function checkUsagePrefs(hasEntry, lastUsed) {
  lastUsed = Math.floor(lastUsed);

  let parentPrefs = runInParent(() => {
    addMessageListener("getPrefValues", () => {
      let prefValues = {};
      prefValues.hasEntry = Services.prefs.getBoolPref(
        "signon.usage.hasEntry",
        false
      );
      prefValues.lastUsed = Services.prefs.getIntPref(
        "signon.usage.lastUsed",
        0
      );
      sendAsyncMessage("prefValues", prefValues);
    });
  });

  parentPrefs.sendAsyncMessage("getPrefValues");
  let prefValues = await new Promise((resolve) => {
    parentPrefs.addMessageListener("prefValues", (values) => {
      parentPrefs.removeMessageListener("prefValues");
      resolve(values);
    });
  })

  info(JSON.stringify(prefValues));

  is(
    prefValues.hasEntry,
    hasEntry,
    "hasEntry usage pref is " + hasEntry
  );
  ok(
    lastUsed - prefValues.lastUsed < 10,
    `lastUsed usage pref (${prefValues.lastUsed}) is within 10 seconds of ${lastUsed}`
  );
}


runChecksAfterCommonInit(startTest);

/** Test for Login Manager: form fill, multiple forms. **/

async function startTest() {
  runInParent(() => {
    Services.prefs.clearUserPref("signon.usage.hasEntry");
    Services.prefs.clearUserPref("signon.usage.lastUsed");
  });

  await checkUsagePrefs(false, 0);
  addLogin();
  await checkUsagePrefs(true, 0);

  let win = window.open("about:blank");
  SimpleTest.registerCleanupFunction(() => win.close());
  await loadFormIntoWindow(DEFAULT_ORIGIN, `
    <form id="form1" action="/">
      <p>This is form 1.</p>
      <input id="username-1" type="text"       name="uname">
      <input id="password-1" type="password"   name="pword">

      <button type="submit" name="submit">Submit</button>
      <button type="reset"> Reset </button>
    </form>`, win);
  await checkLoginFormInFrame(win, "username-1", "testuser", "password-1", "testpass");

  SimpleTest.finish();
}
</script>

<p id="display"></p>

<div id="content" style="display: none">


</div>

<pre id="test"></pre>
</body>
</html>
